graphics refactor and cleanup cleanup of errors and warnings upgrade of some dependencies improved linting
