+++ client/src/controllers/startControls.ts
@@ -1,16 +1,7 @@
 import { World } from "box2d";
-import { registerEars } from "~/audio/sounds";
-import { rigToGamePad } from "~/helpers/utils/gamePad";
-import { rigToKeyboard } from "~/input/getKeyboardInput";
 import { Box2DPreviewMesh } from "~/meshes/Box2DPreviewMesh";
-import { registerHUD } from "~/physics/managers/hudManager";
 import SimpleGUIOverlay from "~/ui/SimpleGUIOverlay";
 import { RayCastConverter } from "~/utils/RayCastConverter";
-
-import { Controller } from "./Controller";
-import GamePadControls from "./GamePadController";
-import MouseKeyboardController from "./MouseKeyboardController";
-import TouchScreenController, { rigToTouchScreen } from "./TouchScreenController";
 
 export function startControls(
 	b2World: World,
@@ -25,9 +16,9 @@ export function startControls(
 	// mcl.register(avatarContactListener);
 	// b2World.SetContactListener(mcl);
 
-	const initiateControls = (controller: Controller) => {
-		registerHUD(gui);
-	};
+	// const initiateControls = (controller: Controller) => {
+	// 	registerHUD(gui);
+	// };
 
 	// rigToGamePad(gamePadAPI => initiateControls(new GamePadControls(gamePadAPI)));
