+++ client/src/factories/ArchitectureModelFactory.ts
@@ -1,30 +1,31 @@
 import { Body } from "box2d";
-import { LoadingManager, Mesh } from "three";
+import { Mesh } from "three";
 import { GLTF, GLTFLoader } from "three/examples/jsm/loaders/GLTFLoader";
 import { cloneMeshByName } from "~/utils/cloneMeshByName";
 
 class ArchitectureModelFactory {
 	private _status: "ready" | "initializing" | "not started" = "not started";
 	private _gltf: GLTF;
-	private _queuedBodies: Map<Body, (wall: Mesh) => void> = new Map();
+	private _bodiesPromisedMeshes: Map<Body, (wall: Mesh) => void> = new Map();
 	private _bodyMeshMap: Map<Body, Mesh> = new Map();
 
-	makeWall(body: Body, onReady: (wall: Mesh) => void) {
-		if (this._status === "ready") {
-			const mesh = this._actuallyMakeWall(body);
-			this._bodyMeshMap.set(body, mesh);
-			onReady(mesh);
-		} else {
-			this._queuedBodies.set(body, onReady);
-			if (this._status === "not started") {
-				this.init(() => {});
+	async requestMesh(body: Body) {
+		return new Promise<Mesh>(resolve => {
+			if (this._status === "ready") {
+				const mesh = this._makeMesh(body);
+				resolve(mesh);
+			} else {
+				this._bodiesPromisedMeshes.set(body, resolve);
+				if (this._status === "not started") {
+					this._init();
+				}
 			}
-		}
+		});
 	}
 
-	unmakeWall(body: Body) {
-		if (this._queuedBodies.has(body)) {
-			this._queuedBodies.delete(body);
+	deleteMesh(body: Body) {
+		if (this._bodiesPromisedMeshes.has(body)) {
+			this._bodiesPromisedMeshes.delete(body);
 		}
 		if (this._bodyMeshMap.has(body)) {
 			const mesh = this._bodyMeshMap.get(body);
@@ -34,49 +35,36 @@ class ArchitectureModelFactory {
 			return undefined;
 		}
 	}
-	init(callback: () => void) {
+	private async _init() {
 		if (this._status === "not started") {
 			this._status = "initializing";
 
-			const loadingManager = new LoadingManager();
-			const gltfLoader = new GLTFLoader(loadingManager);
+			const gltfLoader = new GLTFLoader();
 
-			gltfLoader.load(
-				"game/models/brick-wall.gltf",
-				gltf => {
-					this._gltf = gltf;
-					for (const child of gltf.scene.children) {
-						child.scale.multiplyScalar(0.2);
-
-						child.traverse(obj => {
-							if (obj instanceof Mesh) {
-								obj.castShadow = true;
-								obj.receiveShadow = true;
-							}
-						});
+			const gltf = await gltfLoader.loadAsync("game/models/columns.glb");
+			this._gltf = gltf;
+			for (const child of gltf.scene.children) {
+				child.traverse(obj => {
+					if (obj instanceof Mesh) {
+						obj.castShadow = true;
+						obj.receiveShadow = true;
 					}
+				});
+			}
 
-					this._status = "ready";
-					this._queuedBodies.forEach((onReady, body) => {
-						onReady(this._actuallyMakeWall(body));
-					});
-				},
-				ev => {
-					/* */
-				},
-				ev => {
-					// debugger;
-				}
-			);
+			this._status = "ready";
+			this._bodiesPromisedMeshes.forEach((resolve, body) => {
+				resolve(this._makeMesh(body));
+			});
 		} else {
 			throw new Error("cannot initialize twice");
 		}
 	}
 
-	private _actuallyMakeWall(body: Body) {
-		const mesh = cloneMeshByName(this._gltf, "brick-wall");
+	private _makeMesh(body: Body) {
+		const mesh = cloneMeshByName(this._gltf, "column1");
 		const pos = body.GetPosition();
-		mesh.position.set(pos.x, 0, -pos.y);
+		mesh.position.set(pos.x, -pos.y, 0);
 		this._bodyMeshMap.set(body, mesh);
 		return mesh;
 	}
