+++ client/src/helpers/scenes/Testb2World.ts
@@ -1,11 +1,11 @@
-import { Body, BodyDef, FixtureDef, PolygonShape, ShapeType, Vec2, World } from "box2d";
+import { Body, Vec2, World } from "box2d";
 import { Camera, Color, Fog, PerspectiveCamera, Scene, WebGLRenderer } from "three";
 import device from "~/device";
 import getKeyboardInput from "~/input/getKeyboardInput";
 import { Box2DPreviewMesh, debugPolygonPhysics } from "~/meshes/Box2DPreviewMesh";
 import BaseContactListener from "~/physics/contact listeners/BaseContactListener";
 import { getBodyEventManager } from "~/physics/managers/bodyEventManager";
-import { processDestructions, queueDestruction } from "~/physics/managers/destructionManager";
+import { processDestructions } from "~/physics/managers/destructionManager";
 import { processHUD } from "~/physics/managers/hudManager";
 import {
 	convertTob2Space,
@@ -13,28 +13,11 @@ import {
 	createImprovedPhysicsCircle,
 	createSensorBox,
 	createStaticBox,
-	queryForSingleArchitectureBody,
-	queryForSingleEnvironmentBlock,
-	WallData
+	queryForSingleArchitectureBody
 } from "~/physics/utils/physicsUtils";
-import { RayCastClosestCallback } from "~/physics/utils/rayCastUtils";
-import {
-	BodyData,
-	FixtureData,
-	packFilter,
-	packTransform,
-	packVec2,
-	PolygonShapeData,
-	ShapeData,
-	unpackFilter,
-	unpackTransform,
-	unpackVec2,
-	WorldData
-} from "~/physics/utils/serialUtils";
 import SimpleGUIOverlay from "~/ui/SimpleGUIOverlay";
 import { KeyboardCodes } from "~/utils/KeyboardCodes";
-import { getUrlColor, getUrlFlag, getUrlParam } from "~/utils/location";
-import { rand } from "~/utils/math";
+import { getUrlColor } from "~/utils/location";
 import { RayCastConverter } from "~/utils/RayCastConverter";
 
 import { startControls } from "../../controllers/startControls";
@@ -47,7 +30,6 @@ let isKeyQDown: boolean = false;
 let isKeyZDown: boolean = false;
 let isKeyXDown: boolean = false;
 let isKeyCDown: boolean = false;
-let isKeyVDown: boolean = false;
 
 export default class Testb2World {
 	autoClear = true;
@@ -86,7 +68,7 @@ export default class Testb2World {
 		mcl.register(new BaseContactListener());
 		this.b2World.SetContactListener(mcl);
 
-		const shootRayClosest = new RayCastClosestCallback();
+		// const shootRayClosest = new RayCastClosestCallback();
 
 		/* The Be All And End All Keyboard Listener (a.k.a. THE BUTTON FUNNELER) */
 		getKeyboardInput().addListener(this.HandleKey);
@@ -100,7 +82,7 @@ export default class Testb2World {
 		createStaticBox(this.b2World, -1, -0.9, 0.2, 0.1);
 		createStaticBox(this.b2World, 1, -0.9, 0.2, 0.1);
 
-		const penaltyLine = createSensorBox(
+		createSensorBox(
 			this.b2World,
 			0,
 			-1.5,
@@ -113,19 +95,7 @@ export default class Testb2World {
 			["architecture"],
 			true
 		);
-		const goalLine = createSensorBox(
-			this.b2World,
-			0,
-			1,
-			10,
-			0.1,
-			undefined,
-			undefined,
-			undefined,
-			["goal"],
-			["architecture"],
-			true
-		);
+		createSensorBox(this.b2World, 0, 1, 10, 0.1, undefined, undefined, undefined, ["goal"], ["architecture"], true);
 
 		// const cm = b2World.GetContactManager();
 		// let contacts = cm.m_contactList;
@@ -195,14 +165,14 @@ export default class Testb2World {
 			// if (isKeyVDown) {
 
 			this.selectedBody = queryForSingleArchitectureBody(b2World, clickedb2Space);
-			if (this.selectedBody) {
-				const fixtList = this.selectedBody.GetFixtureList();
-				// console.log("occupyingBlock exists");
-				// console.log(this.selectedBody);
-			} else {
-				this.selectedBody = undefined;
-				// console.log(this.selectedBody);
-			}
+			// if (this.selectedBody) {
+			// 	// const fixtList = this.selectedBody.GetFixtureList();
+			// 	// console.log("occupyingBlock exists");
+			// 	// console.log(this.selectedBody);
+			// } else {
+			// 	this.selectedBody = undefined;
+			// 	// console.log(this.selectedBody);
+			// }
 
 			// }
 
@@ -218,7 +188,7 @@ export default class Testb2World {
 
 		const onDebugMouseMove = (mouseMove: MouseEvent) => {
 			if (this.selectedBody) {
-				this.selectedBody.SetLinearVelocity(new Vec2(0, 0));
+				this.selectedBody.SetLinearVelocity(new Vec2(0.0001, 0));
 				// const position: Vec2 = this.selectedBody.GetPosition()
 				const mouseInb2Space: Vec2 = this.rayCastConverter!(mouseMove.clientX, mouseMove.clientY);
 				this.selectedBody.SetPosition(mouseInb2Space);
@@ -243,9 +213,6 @@ export default class Testb2World {
 				break;
 			case "KeyC":
 				isKeyCDown = down;
-				break;
-			case "KeyV":
-				isKeyVDown = down;
 				break;
 			default:
 				//console.log(code);
@@ -299,149 +266,4 @@ export default class Testb2World {
 		this.camera = camera;
 		this.bgColor = bgColor;
 	}
-}
-
-function loadLevelData(levelDataString: string) {
-	const levelData = JSON.parse(levelDataString) as WorldData;
-	// debugger;
-	for (const bodyData of levelData.bodies) {
-		const bodyDef = new BodyDef();
-		bodyDef.userData = new WallData();
-		bodyDef.angularDamping = bodyData.angularDamping;
-		bodyDef.angularVelocity = bodyData.angularVelocity;
-		bodyDef.type = bodyData.bodyType;
-		bodyDef.fixedRotation = bodyData.fixedRotationFlag;
-		bodyDef.linearDamping = bodyData.linearDamping;
-
-		const body = getBodyEventManager().createBody(bodyDef);
-
-		body.SetLinearVelocity(unpackVec2(bodyData.linearVelocity));
-		body.SetAngularVelocity(bodyData.angularVelocity);
-
-		body.SetTransform(unpackTransform(bodyData.xf));
-		// body.SetTransform(unpackTransform(bodyData.xf0));
-		for (const fixtureData of bodyData.fixtures) {
-			const fixtureDef = new FixtureDef();
-			fixtureDef.density = fixtureData.density;
-
-			fixtureDef.friction = fixtureData.friction;
-			fixtureDef.isSensor = fixtureData.isSensor;
-			fixtureDef.restitution = fixtureData.restitution;
-			fixtureDef.restitutionThreshold = fixtureData.restitutionThreshold;
-
-			switch (fixtureData.shape.type) {
-				case ShapeType.e_chainShape:
-					// debugger;
-					break;
-				case ShapeType.e_circleShape:
-					// debugger;
-					break;
-				case ShapeType.e_edgeShape:
-					// debugger;
-					break;
-				case ShapeType.e_polygonShape:
-					const polyShapeData = fixtureData.shape as PolygonShapeData;
-					const polyShape = new PolygonShape();
-
-					polyShape.m_count = polyShapeData.count;
-					polyShape.m_radius = polyShapeData.radius;
-					polyShape.m_normals = polyShapeData.normals.map(unpackVec2);
-					// const unpackedVertices: Vec2[] = [];
-					// for (const vertice of polyShapeData.vertices) {
-					// 	unpackedVertices.push(unpackVec2(vertice));
-					// } USE MAP TO ITERATE THROUGH ELEMENTS WITH FUNCTION INTO NEW ARRAY WHICH IS THEN ASSIGNED
-					polyShape.m_vertices = polyShapeData.vertices.map(unpackVec2);
-
-					fixtureDef.shape = polyShape;
-
-					// debugger;
-					break;
-				case ShapeType.e_shapeTypeCount:
-					// debugger;
-					break;
-				case ShapeType.e_unknown:
-					// debugger;
-					break;
-			}
-			// debugger;
-			const fixture = body.CreateFixture(fixtureDef);
-			fixture.SetFilterData(unpackFilter(fixtureData.filter));
-			// body.SetUserData({ serializable: true });
-		}
-	}
-}
-
-function saveLevelBeforeUnload(b2World: World) {
-	window.onbeforeunload = ev => {
-		const bodies: BodyData[] = [];
-		const data: WorldData = {
-			bodies
-		};
-
-		let body = b2World.GetBodyList();
-		// debugger;
-		while (body) {
-			if (!(body.m_userData instanceof WallData) || !body.GetFixtureList()) {
-				// if body is not of WallData OR !!!there is no FixtureList!!!
-				// found a bug where Wall bodies existed without their fixtures
-				// (meaning no physics but leaves behind visual artifact in graphics3D test)
-				body = body.m_next;
-				continue;
-			}
-
-			//const pos = packVec2(body.GetPosition());
-			const bodyType = body.m_type;
-			let fixture = body.GetFixtureList();
-			const fixtures: FixtureData[] = [];
-
-			while (fixture) {
-				const shapeSRC = fixture.m_shape;
-				let shape: ShapeData | undefined;
-
-				if (shapeSRC instanceof PolygonShape) {
-					const poly: PolygonShapeData = {
-						type: shapeSRC.m_type,
-						vertices: shapeSRC.m_vertices.map(packVec2),
-						normals: shapeSRC.m_normals.map(packVec2),
-						radius: shapeSRC.m_radius,
-						count: shapeSRC.m_count
-					};
-
-					shape = poly;
-				}
-
-				if (shape) {
-					const fixtureData: FixtureData = {
-						shape,
-						density: fixture.m_density,
-						filter: packFilter(fixture.m_filter),
-						friction: fixture.m_friction,
-						isSensor: fixture.m_isSensor,
-						restitution: fixture.m_restitution,
-						restitutionThreshold: fixture.m_restitutionThreshold
-					};
-					fixtures.push(fixtureData);
-				}
-
-				fixture = fixture.m_next;
-			}
-
-			const bodyData: BodyData = {
-				bodyType,
-				fixtures,
-				angularDamping: body.m_angularDamping,
-				angularVelocity: body.m_angularVelocity,
-				fixedRotationFlag: body.m_fixedRotationFlag,
-				linearDamping: body.m_linearDamping,
-				linearVelocity: packVec2(body.m_linearVelocity),
-				torque: body.m_torque,
-				xf: packTransform(body.m_xf),
-				xf0: packTransform(body.m_xf0)
-			};
-			bodies.push(bodyData);
-			body = body.m_next;
-		}
-
-		localStorage.setItem("level", JSON.stringify(data));
-	};
 }
