+++ client/src/physics/contact listeners/BaseContactListener.ts
@@ -1,8 +1,5 @@
-import { Body, Contact, ContactListener, Fixture, Vec2 } from "box2d";
-import { playWorldSound } from "~/audio/sounds";
+import { Contact, ContactListener, Fixture } from "box2d";
 import { translateCategoryBitsToString } from "~/physics/utils/physicsUtils";
-import { removeFromArray } from "~/utils/arrayUtils";
-import { taskTimer } from "~/utils/taskTimer";
 
 import { queueDestruction } from "../managers/destructionManager";
 
@@ -33,6 +30,9 @@ export default class BaseContactListener extends ContactListener {
 		const tcbts = translateCategoryBitsToString;
 		const fixtA_category = tcbts(fixtureA.m_filter.categoryBits);
 		const fixtB_category = tcbts(fixtureB.m_filter.categoryBits);
+		if (fixtA_category === fixtB_category) {
+			//
+		}
 	}
 }
 function architectureHitsPenalty(architectureFixt: Fixture, penaltyFixt: Fixture) {
