+++ client/src/physics/managers/hudManager.ts
@@ -1,16 +1,12 @@
 import { Body } from "box2d";
-import { Color, Mesh, Scene, Vector3 } from "three";
-import { canvas } from "~/renderer";
-import TextMesh from "~/text/TextMesh";
-import SimpleGUIOverlay, { ToggleButtonUserData } from "~/ui/SimpleGUIOverlay";
+import SimpleGUIOverlay from "~/ui/SimpleGUIOverlay";
 import { removeFromArray } from "~/utils/arrayUtils";
 
 const HUDBodies: Body[] = [];
-let gui: SimpleGUIOverlay | undefined;
+// let gui: SimpleGUIOverlay | undefined;
 
 export async function registerHUD(passedGUI: SimpleGUIOverlay) {
-	gui = passedGUI;
-
+	// gui = passedGUI;
 	// await initializeFullscreenButton();
 }
 
@@ -19,13 +15,13 @@ export function unregisterUserHUD(playerBody: Body) {
 }
 
 export function processHUD(dt: number) {
-	for (const player of HUDBodies) {
-	}
+	// for (const player of HUDBodies) {
+	// }
 }
 
-async function initializeFullscreenButton() {
-	const fullscreenEnterIcon = await gui!.makeFullscreenIcon(
-		window.innerWidth - gui!._relativeWidthButtonSpacing,
-		gui!._relativeHeightButtonSpacing / 2
-	);
-}
+// async function initializeFullscreenButton() {
+// 	return gui!.makeFullscreenIcon(
+// 		window.innerWidth - gui!._relativeWidthButtonSpacing,
+// 		gui!._relativeHeightButtonSpacing / 2
+// 	);
+// }
