+++ client/src/text/TextMesh.ts
@@ -3,8 +3,8 @@ import {
 	BufferGeometry,
 	Camera,
 	Color,
-	Geometry,
 	Group,
+	IUniform,
 	Material,
 	Matrix4,
 	Mesh,
@@ -240,7 +240,7 @@ export default class TextMesh extends Mesh {
 		renderer: WebGLRenderer,
 		scene: Scene,
 		camera: Camera,
-		geometry: Geometry | BufferGeometry,
+		geometry: BufferGeometry,
 		material: Material,
 		group: Group
 	) => {
@@ -297,7 +297,10 @@ export default class TextMesh extends Mesh {
 	}
 
 	private updateMeasurements() {
-		const bb = this.geometry.boundingBox;
+		if (!this.geometry.boundingBox) {
+			this.geometry.computeBoundingBox();
+		}
+		const bb = this.geometry.boundingBox!;
 		this.width = bb.max.x - bb.min.x;
 		this.height = Math.abs(bb.max.y - bb.min.y);
 		this.userData.resolution = new Vector2(this.width, this.height);
@@ -308,16 +311,17 @@ export default class TextMesh extends Mesh {
 }
 
 interface MSDFShaderUniforms {
+	[key: string]: IUniform;
 	msdf: { value: Texture };
 	alphaTest: { value: number };
 	strokeWidth: { value: number };
 	strokeBias: { value: number };
 	strokeColor: { value: Color };
 	opacity: { value: number };
-	clipSpacePosition?: Uniform;
-	pixelSizeInClipSpace?: Uniform;
-	offset?: Uniform;
-	prescale?: Uniform;
+	// clipSpacePosition?: Uniform;
+	// pixelSizeInClipSpace?: Uniform;
+	// offset?: Uniform;
+	// prescale?: Uniform;
 }
 
 const initMaterial = (settings: TextSettings) => {
@@ -476,7 +480,7 @@ const createTextGeometry = (
 	const y = settings.bakedOffset ? settings.bakedOffset.y : 0;
 
 	geometry.computeBoundingBox();
-	const bb = geometry.boundingBox;
+	const bb = geometry.boundingBox!;
 	if (settings.width) {
 		const layoutWidth = geometry.layout.width;
 		bb.max.x = layoutWidth - bb.min.x;
